+++ .idea/misc.xml
@@ -5,18 +5,23 @@
       <map>
         <entry key="..\:/Users/s1/first_app/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.378125" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/activity_image_detail.xml" value="0.152" />
-        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/activity_main.xml" value="0.3333333333333333" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/activity_main.xml" value="0.176" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/activity_post_list.xml" value="0.23136246786632392" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/activity_sign_up.xml" value="0.264" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/activity_splash.xml" value="0.1757852077001013" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/content_image_detail.xml" value="0.4" />
-        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/fragment_free.xml" value="0.176" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/fragment_free.xml" value="0.36342042755344417" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/fragment_gallery.xml" value="0.264" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/fragment_phone_book.xml" value="0.3619791666666667" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/fragment_post_list.xml" value="0.30434782608695654" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/fragment_write_post.xml" value="0.11296859169199595" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/layout.xml" value="0.3463541666666667" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/multi_image_item.xml" value="0.3463541666666667" />
-        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/post_item.xml" value="0.24010416666666667" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/layout/post_item.xml" value="0.264" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/menu/navdrawer_menu.xml" value="0.536" />
         <entry key="..\:/Users/s1/first_app/app/src/main/res/menu/navi_menu.xml" value="0.264" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/menu/options_menu.xml" value="0.18229166666666666" />
+        <entry key="..\:/Users/s1/first_app/app/src/main/res/menu/winner_menu.xml" value="0.18229166666666666" />
       </map>
     </option>
   </component>
