--- app/src/main/java/com/example/first_app/PostListActivity.kt
@@ -1,13 +1,0 @@
-package com.example.first_app
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class PostListActivity : AppCompatActivity() {
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_post_list)
-    }
-}
