+++ app/src/main/res/layout/activity_image_detail.xml
@@ -4,20 +4,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ImageDetailActivity">
 
-    <FrameLayout android:layout_width="0dp"
+    <ImageView
+        android:id="@+id/img_detail"
+        android:layout_width="200dp"
         android:layout_height="0dp"
-        android:id="@+id/fl_container"
+        android:layout_marginTop="20dp"
+        app:layout_constraintDimensionRatio="1:1"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/bnv_main"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"/>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView android:layout_width="0dp"
-        android:layout_height="?attr/actionBarSize" android:id="@+id/bnv_main" android:background="#FFFFFF"
-        app:menu="@menu/navi_menu" app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent" app:layout_constraintRight_toRightOf="parent"/>
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
