+++ app/src/main/res/layout/fragment_write_post.xml
@@ -12,10 +12,12 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:layout_margin="20dp">
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Board" />
+            android:text="Board"
+            android:textColor="#000000" />
 
         <EditText
             android:id="@+id/et_write"
@@ -27,12 +29,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="글쓰기" />
-        <ListView
-            android:id="@+id/lv_post"
-            android:layout_width="409dp"
-            android:layout_height="700dp"
-            tools:layout_editor_absoluteX="1dp"
-            tools:layout_editor_absoluteY="1dp" />
+
     </LinearLayout>
