+++ app/src/main/res/layout/post_item.xml
@@ -1,12 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
+    android:layout_height="match_parent"
+    android:orientation="horizontal"
+    android:layout_marginVertical="10dp"
+    >
+    <ImageView
+        android:id="@+id/iv_post_image"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_weight="1"
+        app:srcCompat="@drawable/monoparental" />
     <TextView
         android:id="@+id/tv_post"
-        android:layout_width="match_parent"
+        android:layout_width="0dp"
         android:layout_height="50dp"
+        android:layout_weight="10"
         />
+
+
 
 </LinearLayout>
